riscv64: auto generate

+++ ccgo_linux_riscv64.go
@@ -36422,7 +36422,7 @@ func _ExprRandFunc(tls *libc.TLS, clientData TClientData, interp uintptr, objc i
 	 * Since the recurrence keeps seed values in the range [1, RAND_IM - 1],
 	 * dividing by RAND_IM yields a double in the range (0, 1).
 	 */
-	dResult = float64((*TInterp)(unsafe.Pointer(iPtr)).FrandSeed) * (libc.Float64FromFloat64(1) / libc.Float64FromInt32(m_RAND_IM))
+	dResult = float64(float64((*TInterp)(unsafe.Pointer(iPtr)).FrandSeed) * (libc.Float64FromFloat64(1) / libc.Float64FromInt32(m_RAND_IM)))
 	/*
 	 * Push a Tcl object with the result.
 	 */
@@ -39248,7 +39248,7 @@ func _BinaryFormatCmd(tls *libc.TLS, dummy TClientData, interp uintptr, objc int
 					if offset%int32(2) != 0 {
 						v49 = *(*uintptr)(unsafe.Pointer(bp + 24))
 						*(*uintptr)(unsafe.Pointer(bp + 24))++
-						*(*uint8)(unsafe.Pointer(v49)) = libc.Uint8FromInt32(value)
+						*(*uint8)(unsafe.Pointer(v49)) = uint8(libc.Uint8FromInt32(value))
 						value = 0
 					}
 					goto _46
@@ -60266,7 +60266,7 @@ func XTcl_TimeObjCmd(tls *libc.TLS, dummy TClientData, interp uintptr, objc int3
 		}
 	}
 	XTcl_GetTime(tls, bp+56)
-	totalMicroSec = float64((*(*TTcl_Time)(unsafe.Pointer(bp + 56))).Fsec-(*(*TTcl_Time)(unsafe.Pointer(bp + 40))).Fsec)*float64(1e+06) + float64((*(*TTcl_Time)(unsafe.Pointer(bp + 56))).Fusec-(*(*TTcl_Time)(unsafe.Pointer(bp + 40))).Fusec)
+	totalMicroSec = float64(float64((*(*TTcl_Time)(unsafe.Pointer(bp + 56))).Fsec-(*(*TTcl_Time)(unsafe.Pointer(bp + 40))).Fsec)*float64(1e+06)) + float64((*(*TTcl_Time)(unsafe.Pointer(bp + 56))).Fusec-(*(*TTcl_Time)(unsafe.Pointer(bp + 40))).Fusec)
 	if *(*int32)(unsafe.Pointer(bp + 32)) <= int32(1) {
 		/*
 		 * Use int obj since we know time is not fractional. [Bug 1202178]
@@ -61247,7 +61247,7 @@ func XTclNRTryObjCmd(tls *libc.TLS, clientData TClientData, interp uintptr, objc
 		(*TTcl_Obj)(unsafe.Pointer((*(*[5]uintptr)(unsafe.Pointer(bp + 16)))[int32(1)])).Fbytes = libc.UintptrFromInt32(0)
 		(*TTcl_Obj)(unsafe.Pointer((*(*[5]uintptr)(unsafe.Pointer(bp + 16)))[int32(1)])).FinternalRep.FlongValue = int64(*(*int32)(unsafe.Pointer(bp + 4)))
 		(*TTcl_Obj)(unsafe.Pointer((*(*[5]uintptr)(unsafe.Pointer(bp + 16)))[int32(1)])).FtypePtr = uintptr(unsafe.Pointer(&XtclIntType)) /* returnCode */
-		if (*(*[5]uintptr)(unsafe.Pointer(bp + 16)))[int32(2)] == libc.UintptrFromInt32(0) {                                              /* errorCodePrefix */
+		if (*(*[5]uintptr)(unsafe.Pointer(bp + 16)))[int32(2)] == libc.UintptrFromInt32(0) {                                               /* errorCodePrefix */
 			if v27 = libc.Bool(libc.UintptrFromInt32(0) == libc.UintptrFromInt32(0)); !v27 {
 				cachePtr2 = (*TInterp)(unsafe.Pointer(libc.UintptrFromInt32(0))).FallocCache
 			}
@@ -165915,7 +165915,7 @@ _20:
 	case int32(m_INST_SUB):
 		dResult = *(*float64)(unsafe.Pointer(bp + 32)) - *(*float64)(unsafe.Pointer(bp + 40))
 	case int32(m_INST_MULT):
-		dResult = *(*float64)(unsafe.Pointer(bp + 32)) * *(*float64)(unsafe.Pointer(bp + 40))
+		dResult = float64(*(*float64)(unsafe.Pointer(bp + 32)) * *(*float64)(unsafe.Pointer(bp + 40)))
 	case int32(m_INST_DIV):
 		/*
 		 * We presume that we are running with zero-divide unmasked if
@@ -172614,7 +172614,7 @@ func XTcl_HashStats(tls *libc.TLS, tablePtr uintptr) (r uintptr) {
 		}
 		tmp = float64(j)
 		if (*TTcl_HashTable)(unsafe.Pointer(tablePtr)).FnumEntries != 0 {
-			average += (tmp + float64(1)) * (tmp / float64((*TTcl_HashTable)(unsafe.Pointer(tablePtr)).FnumEntries)) / float64(2)
+			average += float64((tmp+float64(1))*(tmp/float64((*TTcl_HashTable)(unsafe.Pointer(tablePtr)).FnumEntries))) / float64(2)
 		}
 		goto _2
 	_2:
@@ -247019,7 +247019,7 @@ func _MakeLowPrecisionDouble(tls *libc.TLS, signum int32, significand TTcl_WideU
 				 * 10**exponent. The product will be correct to within 1/2 ulp
 				 * without special handling.
 				 */
-				retval = float64(libc.Int64FromUint64(significand)) * _pow10vals[exponent]
+				retval = float64(float64(libc.Int64FromUint64(significand)) * _pow10vals[exponent])
 				goto returnValue
 			} else {
 				diff = int32(m_QUICK_MAX) - numSigDigs
@@ -247030,8 +247030,8 @@ func _MakeLowPrecisionDouble(tls *libc.TLS, signum int32, significand TTcl_WideU
 					 * significand*10**diff, so we can still compute the value
 					 * with only one roundoff.
 					 */
-					factor = float64(libc.Int64FromUint64(significand)) * _pow10vals[diff]
-					retval = factor * _pow10vals[exponent-int64(diff)]
+					factor = float64(float64(libc.Int64FromUint64(significand)) * _pow10vals[diff])
+					retval = float64(factor * _pow10vals[exponent-int64(diff)])
 					goto returnValue
 				}
 			}
@@ -247730,7 +247730,7 @@ func _ApproximateLog10(tls *libc.TLS, bw TTcl_WideUInt, be int32, bbits int32) (
 		Fword0 int32
 	})(unsafe.Pointer(bp))).Fword0 |= libc.Int32FromInt32(m_EXPONENT_BIAS) << libc.Int32FromInt32(m_EXP_SHIFT)
 	i = be + bbits - int32(1)
-	ds = (*(*float64)(unsafe.Pointer(bp))-float64(1.5))*float64(m_TWO_OVER_3LOG10) + float64(m_LOG10_3HALVES_PLUS_FUDGE) + float64(m_LOG10_2)*float64(i)
+	ds = float64((*(*float64)(unsafe.Pointer(bp))-float64(1.5))*float64(m_TWO_OVER_3LOG10)) + float64(m_LOG10_3HALVES_PLUS_FUDGE) + float64(float64(m_LOG10_2)*float64(i))
 	k = int32(ds)
 	if float64(k) > ds {
 		k--
@@ -248190,13 +248190,13 @@ func _QuickConversion(tls *libc.TLS, _e float64, k int32, k_check int32, flags i
 		}
 		ilim = ilim1
 		k--
-		d = d * float64(10)
+		d = float64(d * float64(10))
 		ieps++
 	}
 	/*
 	 * Compute estimated roundoff error.
 	 */
-	*(*float64)(unsafe.Pointer(bp + 8)) = float64(ieps)*d + float64(7)
+	*(*float64)(unsafe.Pointer(bp + 8)) = float64(float64(ieps)*d) + float64(7)
 	(*(*struct {
 		Fword1 int32
 		Fword0 int32
@@ -249657,7 +249657,7 @@ func XTclInitDoubleConversion(tls *libc.TLS) {
 	/*
 	 * Initialize table of powers of 10 expressed as wide integers.
 	 */
-	_maxpow10_wide = int32(libc.Xfloor(tls, float64(libc.Uint64FromInt64(8)*libc.Uint64FromInt32(m_CHAR_BIT))*libc.Xlog(tls, float64(2))/libc.Xlog(tls, float64(10))))
+	_maxpow10_wide = int32(libc.Xfloor(tls, float64(float64(libc.Uint64FromInt64(8)*libc.Uint64FromInt32(m_CHAR_BIT))*libc.Xlog(tls, float64(2)))/libc.Xlog(tls, float64(10))))
 	_pow10_wide = XTcl_Alloc(tls, uint32(libc.Uint64FromInt32(_maxpow10_wide+libc.Int32FromInt32(1))*libc.Uint64FromInt64(8)))
 	u = uint64(1)
 	i = 0
@@ -249687,7 +249687,7 @@ func XTclInitDoubleConversion(tls *libc.TLS) {
 	 * Initialize a table of powers of ten that can be exactly represented in
 	 * a double.
 	 */
-	x = int32(libc.Float64FromInt32(m_DBL_MANT_DIG) * libc.Xlog(tls, libc.Float64FromInt32(m_FLT_RADIX)) / libc.Xlog(tls, float64(5)))
+	x = int32(float64(libc.Float64FromInt32(m_DBL_MANT_DIG)*libc.Xlog(tls, libc.Float64FromInt32(m_FLT_RADIX))) / libc.Xlog(tls, float64(5)))
 	if x < int32(m_MAXPOW) {
 		_mmaxpow = x
 	} else {
@@ -249750,9 +249750,9 @@ func XTclInitDoubleConversion(tls *libc.TLS) {
 	 * that differs from zero, and the number of mp_digits needed to represent
 	 * the significand of a double.
 	 */
-	_maxDigits = int32((libc.Float64FromInt32(m_DBL_MAX_EXP)*libc.Xlog(tls, libc.Float64FromInt32(m_FLT_RADIX)) + libc.Float64FromFloat64(0.5)*libc.Xlog(tls, float64(10))) / libc.Xlog(tls, float64(10)))
-	_minDigits = int32(libc.Xfloor(tls, float64(-libc.Int32FromInt32(1021)-libc.Int32FromInt32(m_DBL_MANT_DIG))*libc.Xlog(tls, libc.Float64FromInt32(m_FLT_RADIX))/libc.Xlog(tls, float64(10))))
-	_log10_DIGIT_MAX = int32(libc.Xfloor(tls, libc.Float64FromInt32(m_MP_DIGIT_BIT)*libc.Xlog(tls, float64(2))/libc.Xlog(tls, float64(10))))
+	_maxDigits = int32((float64(libc.Float64FromInt32(m_DBL_MAX_EXP)*libc.Xlog(tls, libc.Float64FromInt32(m_FLT_RADIX))) + float64(libc.Float64FromFloat64(0.5)*libc.Xlog(tls, float64(10)))) / libc.Xlog(tls, float64(10)))
+	_minDigits = int32(libc.Xfloor(tls, float64(float64(-libc.Int32FromInt32(1021)-libc.Int32FromInt32(m_DBL_MANT_DIG))*libc.Xlog(tls, libc.Float64FromInt32(m_FLT_RADIX)))/libc.Xlog(tls, float64(10))))
+	_log10_DIGIT_MAX = int32(libc.Xfloor(tls, float64(libc.Float64FromInt32(m_MP_DIGIT_BIT)*libc.Xlog(tls, float64(2)))/libc.Xlog(tls, float64(10))))
 	/*
 	 * Nokia 770's software-emulated floating point is "middle endian": the
 	 * bytes within a 32-bit word are little-endian (like the native
@@ -250251,7 +250251,7 @@ func _Pow10TimesFrExp(tls *libc.TLS, exponent int32, fraction float64, machexp u
 		/*
 		 * Multiply by 10**exponent.
 		 */
-		retval = libc.Xfrexp(tls, retval*_pow10vals[exponent&int32(0xF)], bp)
+		retval = libc.Xfrexp(tls, float64(retval*_pow10vals[exponent&int32(0xF)]), bp)
 		expt += *(*int32)(unsafe.Pointer(bp))
 		i = int32(4)
 		for {
@@ -250259,7 +250259,7 @@ func _Pow10TimesFrExp(tls *libc.TLS, exponent int32, fraction float64, machexp u
 				break
 			}
 			if exponent&(int32(1)<<i) != 0 {
-				retval = libc.Xfrexp(tls, retval*_pow_10_2_n[i], bp)
+				retval = libc.Xfrexp(tls, float64(retval*_pow_10_2_n[i]), bp)
 				expt += *(*int32)(unsafe.Pointer(bp))
 			}
 			goto _1
@@ -250320,7 +250320,7 @@ func _SafeLdExp(tls *libc.TLS, fract float64, expt int32) (r float64) {
 	if expt < minexpt {
 		a = libc.Xldexp(tls, fract, expt-_mantBits-minexpt)
 		b = libc.Xldexp(tls, float64(1), _mantBits+minexpt)
-		retval = a * b
+		retval = float64(a * b)
 	} else {
 		retval = libc.Xldexp(tls, fract, expt)
 	}
